Changed: Moved namespace of split_blocks into transforms/standard
