+++ projects/dxt-lossless-transform-bc3/benches/split_blocks/portable32.rs
@@ -1,5 +1,5 @@
 use criterion::{black_box, BenchmarkId};
-use dxt_lossless_transform_bc3::split_blocks::split::*;
+use dxt_lossless_transform_bc3::transforms::standard::transform::*;
 use safe_allocator_api::RawAlloc;
 
 fn bench_portable32(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
