+++ projects/dxt-lossless-transform-bc3/src/transforms/mod.rs
@@ -1,0 +1,1 @@
+pub mod standard;
