+++ projects/dxt-lossless-transform-bc3/src/transforms/standard/transform/avx512.rs
