+++ projects/dxt-lossless-transform-bc3/src/transforms/standard/transform/portable32.rs
