+++ projects/dxt-lossless-transform-bc3/src/transforms/standard/untransform/avx512.rs
@@ -3,7 +3,7 @@ use core::arch::x86::*;
 #[cfg(target_arch = "x86_64")]
 use core::arch::x86_64::*;
 
-use crate::split_blocks::unsplit::portable::u32_detransform_with_separate_pointers;
+use crate::transforms::standard::untransform::portable::u32_detransform_with_separate_pointers;
 
 /// # Safety
 ///
