+++ projects/dxt-lossless-transform-bc3/src/transforms/standard/untransform/portable.rs
